Move a few things and remove unnecessary directives
