+++ src/Volvox.Helios.Core/Modules/Common/IDocumented.cs
@@ -4,7 +4,7 @@
     {
         string Name { get; }
         string Version { get; }
-        string Synopsis { get; }
+        string Description { get; }
         ReleaseState ReleaseState { get; }
     }
 }
