+++ src/Volvox.Helios.Core/Utilities/metadata.json
@@ -2,13 +2,13 @@
   {
     "Name": "StreamAnnouncerModule",
     "Version": "0.0.1",
-    "Synopsis": "??!?",
+    "Description": "??!?",
     "ReleaseState": 0
   },
   {
     "Name": "StreamerRoleModule",
     "Version": "0.0.1",
-    "Synopsis": "??!?",
+    "Description": "??!?",
     "ReleaseState": 0
   }
 ]
